requirements.txt
discord.py==1.3.3
terminaltables==3.1.0
gspread==3.5.0
oauth2client==4.1.3
